[Faker Library] Fix decimal default generator
now it does not try to generate value out of range

+++ FakerLibrary/DefaultGenerators/DecimalGenerator.cs
@@ -5,6 +5,6 @@ namespace University.DotnetLabs.Lab2.FakerLibrary.DefualtGenerators;
 public sealed class DecimalGenerator : Generator
 {
     Random random = new Random();
-    public override object Generate() => Convert.ToDecimal(random.Next(0,  Convert.ToInt32(decimal.MaxValue)));
+    public override object Generate() => Convert.ToDecimal(random.Next(0,  Convert.ToInt32(int.MaxValue)));
     public DecimalGenerator() => GeneratingType = typeof(decimal);
 }
